Refactored the monolith
