+++ src/routes/login.tsx
@@ -11,7 +11,7 @@ import {
 import { useMutation } from "@tanstack/react-query";
 import { createFileRoute, Navigate } from "@tanstack/react-router";
 import { SubmitHandler, useForm } from "react-hook-form";
-import { login, UnauthorizedError } from "../api";
+import { login, UnauthorizedError } from "../api/api";
 import { useAuthentication } from "../contexts/authentication";
 
 type SearchParams = {
@@ -41,7 +41,7 @@ export const LoginPage: React.FC = () => {
     mutationFn: (data: Inputs) => login(data.username, data.password),
     onSuccess: ({ jwt }) => {
       authenticate(jwt);
-    }
+    },
   });
   const { register, handleSubmit } = useForm<Inputs>();
   const onSubmit: SubmitHandler<Inputs> = async (data) => {
@@ -49,7 +49,7 @@ export const LoginPage: React.FC = () => {
   };
 
   if (state.isAuthenticated) {
-    return <Navigate to={redirect ?? '/'} />;
+    return <Navigate to={redirect ?? "/"} />;
   }
 
   return (
@@ -115,8 +115,9 @@ export const LoginPage: React.FC = () => {
 export const Route = createFileRoute("/login")({
   validateSearch: (search): SearchParams => {
     return {
-      redirect: typeof search.redirect === "string" ? search.redirect : undefined,
-    }
+      redirect:
+        typeof search.redirect === "string" ? search.redirect : undefined,
+    };
   },
   component: LoginPage,
 });
